Move try_to_int to extra header
