+++ test/Jamfile.v2
@@ -37,6 +37,7 @@ run test_catalog.cpp ;
 run test_encoding.cpp ;
 run test_utf.cpp ;
 run test_util.cpp test_helpers.cpp ;
+run test_util_numeric_convert.cpp ;
 run test_date_time.cpp ;
 run test_ios_info.cpp ;
 run test_ios_prop.cpp ;
